Extract module pyk.prelude.utils
